pyproject.toml
[project]
name = "pyapi"
version = "0.1.0"
description = "PyAPI template for a Python API project"
requires-python = ">=3.13"
readme = "README.md"
license = {file = "LICENSE.md"}

authors = [
    { name = "SamWarden", email = "SamWardenSad@gmail.com" },
]
maintainers = [
    { name = "SamWarden", email = "SamWardenSad@gmail.com" },
]

dependencies = [
    "dishka>=1.7.1",
    "environs>=14.3.0",
    "fastapi>=0.117.1",
    "orjson>=3.11.3",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uuid6>=2025.0.1",
    "uvicorn>=0.36.0",
]

[dependency-groups]
dev = ["pre-commit==4.3.*"]
test = ["pytest==8.4.*"]
lint = [
    "mypy==1.18.*",
    "ruff==0.13.*",
]

[project.scripts]
pyapi = "pyapi.__main__:main"

[project.urls]
Homepage = "https://github.com/SamWarden/pyapi"
Repository = "https://github.com/SamWarden/pyapi"
Issues = "https://github.com/SamWarden/pyapi/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools==80.9.0"]
build-backend = "setuptools.build_meta"
